encode & decode form
I create forms with API calls where we can choose image and song to encode or decode. I have to implement how to get the output file but I need that the API is working. Furthermore, you can choose the local port you use for the API, because some people doesn't use the default port to run the API.